优化兼容性

+++ pref/src/common.rs
@@ -135,12 +135,12 @@ impl GenSysStat {
 
     //获取当前进程id
     pub fn current_pid(&self) -> usize {
-        sysinfo::get_current_pid()
+        sysinfo::get_current_pid() as usize
     }
 
     //获取当前进程的基础状态
     pub fn current_process_info(&self) -> (usize, String, PathBuf, Vec<String>, f32, u64, u64, ProcessState) {
-        let pid = sysinfo::get_current_pid();
+        let pid = sysinfo::get_current_pid() as usize;
         self.inner.borrow_mut().refresh_process(pid);
 
         let inner = self.inner.borrow();
